+++ ft_strmapi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: min-yeongjae <min-yeongjae@student.42.f    +#+  +:+       +#+        */
+/*   By: yeomin <yeomin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/08 14:49:30 by min-yeongja       #+#    #+#             */
-/*   Updated: 2023/03/15 00:00:25 by min-yeongja      ###   ########.fr       */
+/*   Created: 2023/03/08 14:49:30 by yeomin            #+#    #+#             */
+/*   Updated: 2023/03/18 13:25:14 by yeomin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
-	char			*new_str;
+	size_t	i;
+	char	*new_str;
 
 	i = 0;
 	new_str = malloc(sizeof(char) *(ft_strlen(s) + 1));
